requirements.txt
docling==2.14.0
docling-core==2.12.1
langchain==0.3.17
langchain-core==0.3.33
langchain-docling==0.2.0
langchain-huggingface==0.1.2
python-dotenv==1.0.1
